+++ app/build.gradle
@@ -88,8 +88,8 @@ dependencies {
     implementation "com.fasterxml.jackson.core:jackson-annotations:2.10.1"
     implementation "javax.validation:validation-api:2.0.1.Final"
     implementation "com.github.filosganga:geogson-core:1.2.21"
-    implementation files('C:\\Users\\ps664\\Documents\\BrAPI\\brapi-java-client-2\\brapi\\brapi-java-model\\target\\brapi-java-model-2.0-SNAPSHOT.jar')
-    implementation files('C:\\Users\\ps664\\Documents\\BrAPI\\brapi-java-client-2\\brapi\\brapi-java-client\\target\\brapi-java-client-2.0-SNAPSHOT.jar')
+    implementation "org.brapi:brapi-java-client:2.0-SNAPSHOT"
+    implementation "org.brapi:brapi-java-model:2.0-SNAPSHOT"
 
     testImplementation 'androidx.test:core:1.3.0'
     testImplementation 'org.robolectric:robolectric:4.3.1'
