+++ app/src/main/java/com/fieldbook/tracker/activities/ConfigActivity.java
@@ -45,10 +45,9 @@ import android.widget.RadioButton;
 import android.widget.Toast;
 
 import com.fieldbook.tracker.R;
-import com.fieldbook.tracker.brapi.BrAPIService;
+import com.fieldbook.tracker.brapi.service.BrAPIService;
 import com.fieldbook.tracker.brapi.BrapiAuthDialog;
 import com.fieldbook.tracker.database.DataHelper;
-import com.fieldbook.tracker.database.Migrator;
 import com.fieldbook.tracker.objects.FieldObject;
 import com.fieldbook.tracker.preferences.GeneralKeys;
 import com.fieldbook.tracker.preferences.PreferencesActivity;
@@ -66,7 +65,6 @@ import com.michaelflisar.changelog.internal.ChangelogDialogFragment;
 
 import java.io.File;
 import java.io.FileWriter;
-import java.io.IOException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Calendar;
