--- app/src/main/java/com/fieldbook/tracker/brapi/BrapiStudySummary.java
@@ -1,23 +1,0 @@
-package com.fieldbook.tracker.brapi;
-
-public class BrapiStudySummary {
-
-    private String studyName;
-    private String studyDbId;
-
-    public String getStudyName() {
-        return studyName;
-    }
-
-    void setStudyName(String studyName) {
-        this.studyName = studyName;
-    }
-
-    String getStudyDbId() {
-        return studyDbId;
-    }
-
-    void setStudyDbId(String studyDbId) {
-        this.studyDbId = studyDbId;
-    }
-}
