+++ app/src/main/java/com/fieldbook/tracker/brapi/model/BrapiObservation.java
@@ -1,19 +1,14 @@
-package com.fieldbook.tracker.brapi;
+package com.fieldbook.tracker.brapi.model;
 
-import android.os.Build;
-
-import androidx.annotation.RequiresApi;
-
-import org.threeten.bp.DateTimeUtils;
 import org.threeten.bp.OffsetDateTime;
 import org.threeten.bp.format.DateTimeFormatter;
 import org.threeten.bp.format.DateTimeParseException;
 
 public class BrapiObservation {
 
-    protected OffsetDateTime timestamp;
-    String unitDbId;
-    String variableDbId;
+    private OffsetDateTime timestamp;
+    private String unitDbId;
+    private String variableDbId;
     private String dbId;
     private OffsetDateTime lastSyncedTime;
     private String fieldbookDbId;
@@ -38,7 +33,6 @@ public class BrapiObservation {
     public OffsetDateTime getTimestamp() {
         return timestamp;
     }
-
 
     public void setTimestamp(OffsetDateTime timestamp) {
         this.timestamp = timestamp;
@@ -104,7 +98,7 @@ public class BrapiObservation {
     }
 
     // Objects.hashCode Jdk7
-    private int arraysHashCode(Object a[]) {
+    private int arraysHashCode(Object[] a) {
         if (a == null)
             return 0;
