+++ app/src/main/java/com/fieldbook/tracker/brapi/model/BrapiProgram.java
@@ -1,4 +1,4 @@
-package com.fieldbook.tracker.brapi;
+package com.fieldbook.tracker.brapi.model;
 
 public class BrapiProgram {
     private String programName;
