+++ app/src/main/java/com/fieldbook/tracker/brapi/model/BrapiTrial.java
@@ -1,4 +1,4 @@
-package com.fieldbook.tracker.brapi;
+package com.fieldbook.tracker.brapi.model;
 
 public class BrapiTrial {
     private String trialName;
