+++ build.gradle
@@ -20,5 +20,19 @@ allprojects {
         google()
         maven { url "https://dl.bintray.com/pheno-apps/pheno-apps-java-client-library" }
         maven { url "https://jitpack.io" }
+
+        maven {
+            url = "https://maven.pkg.github.com/Breeding-Insight/brapi"
+            credentials {
+                username = getLocalProp("GITHUB_USERNAME")
+                password = getLocalProp("GITHUB_TOKEN")
+            }
+        }
     }
+}
+
+static def getLocalProp(key) {
+    def Properties props = new Properties()
+    props.load(new FileInputStream(new File('local.properties')))
+    return props[key]
 }
